pyproject.toml
[tool.poetry]
name = "aisee"
version = "0.0.1"
description = "aisee is a Python package for image classification using PyTorch and timm created by Instituto de Ingeniería del Conocimiento"
authors = ["IIC"]
license = "MIT"
readme = "README.md"
homepage = "https://aisee.readthedocs.io"
repository = "https://github.com/iiconocimiento/aisee"
documentation = "https://aisee.readthedocs.io"
keywords=["aisee", "computer vision", "vision", "timm", "pytorch"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering",
]

[tool.poetry.dependencies]
python = ">=3.9, <3.12"
datasets = ">= 2"
huggingface-hub = ">= 0.13"
pandas = ">= 1"
Pillow = ">= 8.4.0"
scikit-learn = ">= 1"
timm = ">= 0.5.4"
torch = ">= 1.13.1"
torchvision = ">= 0.11.3"

[tool.poetry.group.dev.dependencies]
pytest = "7.3.1"
pytest-cov = "3.0.0"
coverage = {extras = ["toml"], version ="6.5.0"}
pytest-ruff = "0.0.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = [
    "--cov-report", "html:cover",
    "--cov-report", "term",
    "--cov-config=pyproject.toml",
    "--junitxml=report.xml",
    "--cov=aisee",
    "--durations=20",
    "--ignore=docs/",
]
filterwarnings = [
    "ignore::PendingDeprecationWarning",
    "ignore::RuntimeWarning",
    "ignore::UserWarning",
]

[tool.coverage.paths]
source = ["aisee/"]

[tool.coverage.report]
omit = [
    "**__init__.py",
    "**test**.py",
    "**/docs/***",
    "**.pyc"
]
show_missing = false
precision = 2

[tool.ruff]
select = [
    "E",
    "F",
    "W",
    "I",
    "N",
    "D",
    "UP",
    "ANN",
    "S",
    "B",
    "A",
    "COM",
    "C4",
    "PIE",
    "T20",
    "RET",
    "SIM",
    "ARG",
    "PTH",
    "ERA",
    "PD",
    "PL",
    "RUF",
]
unfixable = ["B007"]
ignore = ["E501", "D100", "D103", "D104", "PLR0913", "PLR0915", "PLR2004", "S101", "ANN101"]

src = ["aisee", "tests"]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "__init__.py",
    "docs/**",
]

# Same as Black.
line-length = 88

target-version = "py39"

[tool.ruff.pydocstyle]
convention = "numpy"

[tool.ruff.isort]
known-first-party = [
    "aisee",
    "VisionClassifier",
    "Trainer",
    "DatasetFromFolder",
    "DatasetFromDataFrame",
    "DatasetFromSingleImage"
]

[tool.ruff.pep8-naming]
ignore-names = ["X"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["E402"]

[tool.ruff.flake8-annotations]
ignore-fully-untyped = true
